// types.ts
export enum Sport {
  NBA = 'NBA',
  NFL = 'NFL',
  BASEBALL = 'BASEBALL',
  HOCKEY = 'HOCKEY',
  SOCCER = 'SOCCER',
  COLLEGE_FOOTBALL = 'COLLEGE_FOOTBALL',
  COLLEGE_BASKETBALL = 'COLLEGE_BASKETBALL',
  WNBA = 'WNBA',
  TENNIS = 'TENNIS',
  GOLF = 'GOLF',
  MMA = 'MMA',
  BASKETBALL = 'BASKETBALL'
}
// ...
export interface League {
  id: string;
  name: string;
  sport: Sport;
  apiEndpoint: string;
  oddsKey?: string;
}